Update CoreAdminConfigurationExtensions.cs FindDbContexts method to fix duplicated DbSets
Hey @edandersen, here is a fix for the duplicated DbSets. I've tested it to work locally in my own project, but if you can test it with the demo that would be much appreciated! :-)

The fix was to ensure checking if the implementation type of discovered services already has been discovered prior to adding the DiscoveredDbContextType to the services.

Fixes #10